Crate syntex_errors

Re-exports

pub use emitter::ColorConfig;

pub use diagnostic::Diagnostic;

pub use diagnostic::SubDiagnostic;

pub use diagnostic::DiagnosticStyledString;

pub use diagnostic::StringPart;

pub use diagnostic_builder::DiagnosticBuilder;

Modules

diagnostic

diagnostic_builder

emitter

registry

snippet

styled_buffer

Structs

CodeSuggestion

ExplicitBug
Signifies that the compiler died with an explicit call to .bug or .span_bug rather than a failed assertion, etc.

FatalError
Used as a return value to signify a fatal error occurred. (It is also used as the argument to panic at the moment, but that will eventually not be true.)

Handler
A handler deals with errors; certain errors (fatal, bug, unimpl) may cause immediate exit, others log errors for later reporting.

Substitution
See the docs on CodeSuggestion::substitutions

Enums

Level

RenderSpan

Traits

CodeMapper

Functions

expect