Struct syntex_errors::diagnostic_builder::DiagnosticBuilder

#[must_use]

pub struct DiagnosticBuilder<'a> { /* fields omitted */ }

Used for emitting structured error messages and other diagnostic information.

Methods

impl<'a> DiagnosticBuilder<'a>

pub fn emit(&mut self)

Emit the diagnostic.

pub fn span_label<T: Into<String>>(&mut self, span: Span, label: T) -> &mut Self

Add a span/label to be included in the resulting snippet. This is pushed onto the MultiSpan that was created when the diagnostic was first built. If you don't call this function at all, and you just supplied a Span to create the diagnostic, then the snippet will just include that Span, which is called the primary span.

pub fn note_expected_found(
&mut self,
label: &Display,
expected: DiagnosticStyledString,
found: DiagnosticStyledString
) -> &mut Self

pub fn note_expected_found_extra(
&mut self,
label: &Display,
expected: DiagnosticStyledString,
found: DiagnosticStyledString,
expected_extra: &Display,
found_extra: &Display
) -> &mut Self

pub fn note(&mut self, msg: &str) -> &mut Self

pub fn span_note<S: Into<MultiSpan>>(&mut self, sp: S, msg: &str) -> &mut Self

pub fn warn(&mut self, msg: &str) -> &mut Self

pub fn span_warn<S: Into<MultiSpan>>(&mut self, sp: S, msg: &str) -> &mut Self

pub fn help(&mut self, msg: &str) -> &mut Self

pub fn span_help<S: Into<MultiSpan>>(&mut self, sp: S, msg: &str) -> &mut Self

pub fn span_suggestion(
&mut self,
sp: Span,
msg: &str,
suggestion: String
) -> &mut Self

pub fn span_suggestions(
&mut self,
sp: Span,
msg: &str,
suggestions: Vec<String>
) -> &mut Self

pub fn set_span<S: Into<MultiSpan>>(&mut self, sp: S) -> &mut Self

pub fn code(&mut self, s: String) -> &mut Self

pub fn new(
handler: &'a Handler,
level: Level,
message: &str
) -> DiagnosticBuilder<'a>

Convenience function for internal use, clients should use one of the struct_* methods on Handler.

pub fn new_with_code(
handler: &'a Handler,
level: Level,
code: Option<String>,
message: &str
) -> DiagnosticBuilder<'a>

Convenience function for internal use, clients should use one of the struct_* methods on Handler.

pub fn into_diagnostic(self) -> Diagnostic