Module syntex_errors::emitter

Structs

BufferedWriter
Buffered writer gives us a way on Unix to buffer up an entire error message before we output it. This helps to prevent interleaving of multiple error messages when multiple compiler processes error simultaneously

EmitterWriter

Enums

ColorConfig

Destination

Constants

MAX_HIGHLIGHT_LINES
maximum number of lines we will print for each error; arbitrary.

MAX_SUGGESTIONS
maximum number of suggestions to be shown

Traits

Emitter
Emitter trait for emitting errors.

Type Definitions

BufferedStderr