Struct syntex_errors::CodeSuggestion

pub struct CodeSuggestion {
    pub substitution_parts: Vec<Substitution>,
    pub msg: String,
}

Fields
substitution_parts: Vec<Substitution>

Each substitute can have multiple variants due to multiple applicable suggestions

foo.bar might be replaced with a.b or x.y by replacing foo and bar on their own:

vec![
    (0..3, vec!["a", "x"]),
    (4..7, vec!["b", "y"]),
]

or by replacing the entire span:

vec![(0..7, vec!["a.b", "x.y"])]

msg: String
Methods

impl CodeSuggestion

pub fn substitution_spans<'a>(
&'a self
) -> Map<<&'a Vec<Substitution> as IntoIterator>::IntoIter, fn(_: &Substitution) -> Span>

Returns the number of substitutions

pub fn splice_lines(&self, cm: &CodeMapper) -> Vec<String>

Returns the assembled code suggestions.